+++ src/types/extension-status.ts
@@ -1,5 +1,0 @@
-type TExtensionStatus = 'EXTENSION_IS_INSTALLED_AND_CONNECTED' |
-                        'EXTENSION_IS_INSTALLED_AND_NOT_CONNECTED' |
-                        'EXTENSION_IS_NOT_INSTALLED'
-
-export default TExtensionStatus