+++ src/types/index.ts
@@ -1,12 +1,9 @@
 import TSemaphoreProof from "./semaphore-proof"
-import TExtensionStatus from './extension-status'
 import TErrorType from "./error-type"
 import TRequestType from "./request-type"
 import TRequest from './request'
 import TResponse from './response'
 import TGenerateSignature from './generate-signature'
-import TTask from "./task"
-import TVerificationData from './verification-data'
 import TScoreMessage from "./score-message"
 import {
   TOutboundMessage,
@@ -19,32 +16,14 @@ import {
   TInboundMessage,
   TProofsResponseMessage
 } from './inbound-message'
-import {
-  TInboundMessageType,
-  TMessage,
-  TOutboundMessageType
-} from './message'
 import TCall3 from './call3'
 import TMode from "./mode"
-export { TTaskGroupCheck } from './task-group-check'
-export { TTaskGroup } from './task-group'
 
-export type { Result } from './result';
-export type { TNotarizationStatus } from './notarization-status';
-export type { TNotarizationStep } from './notarization-step';
-export type { TNotarizationGroup } from './notarization-group';
-export type { TNotarizationGroupCheck } from './notarization-group-check';
-export type { TConnectionQuality } from './connection-quality';
-export type { TSemaphoreProofServer } from './semaphore-proof-server';
-export type { TUserStatus } from './user-status';
-
-export type { TUser } from './user';
-export type { TVerificationStatus } from './verification-status';
-export type { TVerification } from './verification';
-
-export type { TTaskServer } from './task-server';
-export type { TVerifyResult } from './verify-result';
-
+export type { TTaskCheck } from './task-check'
+export type { TTaskGroup } from './task-group'
+export type { TTask } from './task'
+export type { TTasksConfig } from './tasks-config'
+export type { TCredentialGroupPointsMap } from './credential-group-points-map'
 
 export {
   TMode,
@@ -55,19 +34,12 @@ export {
   TCloseModalMessage,
   TGenerateUserKeyMessage,
   TInboundMessage,
-
-  TInboundMessageType,
-  TMessage,
-  TOutboundMessageType,
   TProofsResponseMessage,
-  TExtensionStatus,
   TErrorType,
   TRequestType,
   TRequest,
   TGenerateSignature,
   TResponse,
-  TTask,
   TCall3,
-  TVerificationData,
   TScoreMessage
 }