+++ src/types/notarization-status.ts
@@ -1,1 +1,0 @@
-export type TNotarizationStatus = 'completed' | 'current' | 'disabled';