+++ src/types/task-check.ts
@@ -1,0 +1,5 @@
+export type TTaskCheck = {
+  key: string
+  type: string
+  value: string
+}