+++ src/types/task-group.ts
@@ -1,8 +1,8 @@
-import { TTaskGroupCheck } from './';
+import { TTaskCheck } from './task-check'
 
 export type TTaskGroup = {
-  points: number;
-  semaphoreGroupId: string;
-  credentialGroupId: string;
-  checks?: TTaskGroupCheck[];
-};
+  points: number
+  semaphoreGroupId: string
+  credentialGroupId: string
+  checks: TTaskCheck[]
+}