+++ src/types/user.tsx
@@ -1,8 +1,0 @@
-export type TUser = {
-  key: string | null;
-  id: string | null;
-  address: string | null;
-  loading: boolean;
-  apiKey: string | null;
-  scope: null | string
-};